add fake users creator, ITableGenerator (no picture)

+++ MyBase.WEB/Models/PageInfo.cs
@@ -12,7 +12,7 @@ namespace MyBase.WEB.Models
         public int TotalItems { get; set; } 
         public int TotalPages 
         {
-            get { return (int)Math.Ceiling((decimal)TotalItems / PageSize); }
+            get { return (int)Math.Ceiling((double)TotalItems / PageSize); }
         }        
     }
 }
